Refactoring our middleware
* inlines our sendImage function
* renames routes/avatars to routes/index
