+++ src/index.ts
@@ -1,4 +1,4 @@
-import routes from './routes/avatars';
+import routes from './routes';
 
 export default routes;
 (module as any).exports = routes;
