--- src/routes/common.ts
@@ -1,9 +1,0 @@
-const common = {
-  sendImage: (err, stdout, req, res, next?) => {
-    res.setHeader('Expires', new Date(Date.now() + 604800000));
-    res.setHeader('Content-Type', 'image/png');
-    stdout.pipe(res);
-  },
-};
-
-export default common;
